requirements.txt
# --- Web Server & API ---
fastapi
uvicorn[standard]

# --- Giao diện (Frontend) ---
streamlit
requests

# --- Admin Panel ---
sqladmin

# --- Database (ORM & Drivers) ---
SQLAlchemy
asyncpg
psycopg-binary

# --- LangChain & LangGraph (Lõi) ---
langchain
langchain-core
langgraph

# --- LangChain (Components) ---
langchain-cohere
langchain-huggingface
langchain-chroma

# --- VectorStore & Embeddings ---
chromadb
sentence-transformers

# --- Utilities ---
python-dotenv
Werkzeug